pyproject.toml
[tool.poetry]
name = "topical-api"
version = "0.1.0"
description = "todo"
authors = ["Britton W. <britton.wolfe1999@gmail.com>", "Steve C. <scarter13@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.5"
gunicorn = "^20.0.4"
psycopg2 = "^2.8.5"
djoser = "^2.0.3"
django-rest-framework = "^0.1.0"
django-environ = "^0.4.5"
django-heroku = "^0.3.1"
django-cors-headers = "^3.4.0"

[tool.poetry.dev-dependencies]
prospector = "^1.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
